pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "metrics-forge"
version = "1.0.0"
description = "Generate Prometheus alerting rules and Grafana dashboards from service definitions"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "click>=8.0.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
metrics-forge = "metrics_forge.cli:main"

[tool.setuptools.packages.find]
include = ["metrics_forge*"]
